moved logic out of server.js into data and routes and fixed console log output
